requirements.txt
Pillow
Flask
prometheus-flask-exporter
requests
grpcio-tools
grpcio
furl
pytest
